First Commit: added two slash commands and general set up for the bot
Added /tldr, takes in one parameter representing how many hours of conversation you'd like to go back to summarize

Added /help, which replies with an Embed describing the bot and its commands.

Started a JSONHandler class which is intended to handle the future JSON to be created for the LLM to analyze.

Next steps: Think of a reasonable way for a JSON to be formatted for conversation, and create a JSON file for an LLM to use as input. If a JSON is even the best way to take it in. Research if there is lighter options.